修改添加纪录的参数 增加查询记录的功能

+++ src/nju/model/po/RecordPO.java
@@ -19,7 +19,7 @@ public class RecordPO implements Serializable{
 	private Calendar date;
 	private float amount;
 	
-	public RecordPO(UserPO user,RecordKind kind,Calendar date,int amount) {
+	public RecordPO(UserPO user,RecordKind kind,Calendar date,float amount) {
 		this.user = user;
 		this.kind = kind;
 		this.date = date;
